+++ tests/run-tests.js
@@ -1,0 +1,2 @@
+import './scoringEngine.test.js';
+import './sessionSetup.test.js';